Move x:Bind control samples to folder
