+++ src/Uno.UI.Tests/Windows_UI_Xaml_Data/xBindTests/Controls/Binding_Control.xaml
@@ -1,4 +1,4 @@
-﻿<Page x:Class="Uno.UI.Tests.Windows_UI_Xaml_Data.xBindTests.Given_xBind_Binding_Control"
+﻿<Page x:Class="Uno.UI.Tests.Windows_UI_Xaml_Data.xBindTests.Controls.Binding_Control"
 	  xmlns:sys="using:System"
 	  xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 	  xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
