+++ src/Uno.UI.Tests/Windows_UI_Xaml_Data/xBindTests/Controls/DataTemplate_Control.xaml
@@ -1,8 +1,8 @@
-﻿<Page x:Class="Uno.UI.Tests.Windows_UI_Xaml_Data.xBindTests.Given_xBind_DataTemplate_Control"
+﻿<Page x:Class="Uno.UI.Tests.Windows_UI_Xaml_Data.xBindTests.Controls.DataTemplate_Control"
 	  xmlns:sys="using:System"
 	  xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 	  xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-	  xmlns:local="using:Uno.UI.Tests.Windows_UI_Xaml_Data.xBindTests"
+	  xmlns:local="using:Uno.UI.Tests.Windows_UI_Xaml_Data.xBindTests.Controls"
 	  xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
 	  xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
 	  mc:Ignorable="d">
