+++ src/Uno.UI.Tests/Windows_UI_Xaml_Data/xBindTests/Controls/StaticResource_Control.xaml
@@ -1,0 +1,14 @@
+﻿<Page x:Class="Uno.UI.Tests.Windows_UI_Xaml_Data.xBindTests.Controls.StaticResource_Control"
+	  xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+	  xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+	  xmlns:local="using:Uno.UI.Tests.Windows_UI_Xaml_Data.xBindTests.Controls"
+	  xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+	  xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+	  mc:Ignorable="d">
+
+	<Page.Resources>
+		<CollectionViewSource x:Name="mySource"
+							  x:FieldModifier="public"
+							  Source="{x:Bind MyProperty}" />
+	</Page.Resources>
+</Page>
