+++ src/Uno.UI.Tests/Windows_UI_Xaml_Data/xBindTests/Controls/StaticResource_Control.xaml.cs
@@ -15,14 +15,14 @@ using Windows.UI.Xaml.Navigation;
 
 // The Blank Page item template is documented at https://go.microsoft.com/fwlink/?LinkId=234238
 
-namespace Uno.UI.Tests.Windows_UI_Xaml_Data.xBindTests
+namespace Uno.UI.Tests.Windows_UI_Xaml_Data.xBindTests.Controls
 {
 	/// <summary>
 	/// An empty page that can be used on its own or navigated to within a Frame.
 	/// </summary>
-	public sealed partial class Given_StaticResource_Control : Page
+	public sealed partial class StaticResource_Control : Page
 	{
-		public Given_StaticResource_Control()
+		public StaticResource_Control()
 		{
 			this.InitializeComponent();
 		}
